Print all version numbers in decimal.
This CL reverts a recent change that makes firmware versions appear in
hexidecimal again when performing a FW update.  This is confusing to have
the version number sometimes in hex and sometimes in decimal if it isn't
preceded by 0x there's no way to tell the difference, so let's just use
decimal to be consistent.

Signed-off-by: Charlie Mooney <charliemooney@chromium.org>